+++ src/main/java/bridge/domain/generator/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain.generator;
 
 @FunctionalInterface
 public interface BridgeNumberGenerator {
