+++ src/main/java/bridge/input/validate/BridgeValidator.java
@@ -1,4 +1,4 @@
-package bridge.input;
+package bridge.input.validate;
 
 import java.util.List;
